+开发下载功能中 ++构建调用内置下载器 ++++下载记录更新，存入数据库 +++下载记录 room +瞎改逻辑

+++ Ti/app/src/main/java/com/example/kiylx/ti/corebase/DownloadInfo.java
@@ -127,7 +127,7 @@ public class DownloadInfo {
     }
 
 
-    public double getTotalLength() {
+    public long getTotalLength() {
         return totalLength;
     }
